requirements.txt
pyrotgfork
tgcrypto==1.2.5
python-dotenv==1.0.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
openai==1.50.2
httpx==0.27.2